refactor(settings): 拆分自定义分辨率实现并统一存储访问
- 新增 CustomResolutionsStore 统一 load/save/add,替换 StreamSettings、GameMenu、PcView 各自的手写解析与排序
- 新增 ResolutionFormat 承载纯逻辑(输入校验、比例化简、识别标签、等比小图尺寸)
- 校验改为类型化 ResolutionInputError(字段枚举 + 原因枚举),文案在渲染层经 stringResource 映射,替换 Pair<Int, String>
- 对话框拆为 CustomResolutionsDialog(窗口与状态)与 CustomResolutionsDialogUi(组件);Resolution 提升为顶层 data class,常量并入 Store
- 修复像素文案单位进位:中文「万像素」按 px/10000、其他语言 MP 按 px/1000000;locale 取自 resources.configuration 与资源解析同源(LocalConfiguration.locales 存在不同步问题)
