+++ app/src/main/java/com/limelight/preferences/StreamSettings.kt
@@ -1106,46 +1106,16 @@ class StreamSettings : AppCompatActivity() {
         }
 
         private fun addCustomResolutionsEntries() {
-            val storage = requireActivity().getSharedPreferences(CustomResolutionsConsts.CUSTOM_RESOLUTIONS_FILE,
-                MODE_PRIVATE
-            )
-            val stored = storage.getStringSet(CustomResolutionsConsts.CUSTOM_RESOLUTIONS_KEY, null)
             val pref = findPreference<ListPreference>(PreferenceConfiguration.RESOLUTION_PREF_STRING)!!
-
             val preferencesList = listOf(*pref.entryValues)
 
-            if (stored.isNullOrEmpty()) {
-                return
-            }
-
-            val lengthComparator = Comparator<String> { s1, s2 ->
-                val s1Size = s1.split("x")
-                val s2Size = s2.split("x")
-
-                val w1 = s1Size[0].toInt()
-                val w2 = s2Size[0].toInt()
-
-                val h1 = s1Size[1].toInt()
-                val h2 = s2Size[1].toInt()
-
-                if (w1 == w2) {
-                    h1.compareTo(h2)
-                } else {
-                    w1.compareTo(w2)
-                }
-            }
-
-            val list = ArrayList(stored)
-            Collections.sort(list, lengthComparator)
-
-            for (storedResolution in list) {
+            for (resolution in CustomResolutionsStore.load(requireActivity())) {
+                val storedResolution = resolution.toString()
                 if (preferencesList.contains(storedResolution)) {
                     continue
                 }
-                val resolution = storedResolution.split("x")
-                val width = resolution[0].toInt()
-                val height = resolution[1].toInt()
-                val aspectRatio = AspectRatioConverter.getAspectRatio(width, height)
+
+                val aspectRatio = AspectRatioConverter.getAspectRatio(resolution.width, resolution.height)
                 var displayText = "Custom "
 
                 if (aspectRatio != null) {
