+++ app/src/main/res/values-zh-rTW/strings.xml
@@ -759,6 +759,7 @@
     <string name="custom_resolution_field_width">寬</string>
     <string name="custom_resolution_field_height">高</string>
     <string name="custom_resolution_readout_hint">輸入後自動顯示畫面比例</string>
+    <!-- 參數按語言進位(見 resolutionPixelsText):中文傳「萬像素」(px/10000),其他語言傳 MP(px/1000000)。翻譯時保持單位與進位一致。 -->
     <string name="custom_resolution_pixels_format">%1$.0f 萬像素</string>
     <string name="custom_resolution_error_empty_width">請輸入寬度</string>
     <string name="custom_resolution_error_empty_height">請輸入高度</string>
