+++ app/src/main/res/values/strings.xml
@@ -284,6 +284,7 @@
     <string name="custom_resolution_field_width">Width</string>
     <string name="custom_resolution_field_height">Height</string>
     <string name="custom_resolution_readout_hint">Aspect ratio appears as you type</string>
+    <!-- Arg is locale-scaled (see resolutionPixelsText): zh receives 万 (px/10000), others receive MP (px/1000000). Keep the unit consistent with the scale. -->
     <string name="custom_resolution_pixels_format">%1$.1f MP</string>
     <string name="custom_resolution_error_empty_width">Enter a width</string>
     <string name="custom_resolution_error_empty_height">Enter a height</string>
